TerraMind-project/backend/api/analyze_land.py
from fastapi import APIRouter, UploadFile, File, HTTPException, Form
from pydantic import BaseModel
import numpy as np
import cv2

from ..models.ndvi_processor import calculate_ndvi
from ..utils.sentinelhub_api import get_satellite_image

router = APIRouter()

class AnalysisRequest(BaseModel):
    lat: float
    lon: float

class AnalysisResult(BaseModel):
    plot_id: int
    ndvi: float
    evi: float
    soil_quality: float
    timestamp: str

@router.post("/analyze_land_coords", response_model=AnalysisResult)
async def analyze_land_from_coordinates(req: AnalysisRequest):
    """Analyzes land from geographic coordinates by fetching a satellite image."""
    try:
        # 1. Fetch image from SentinelHub
        image_data = await get_satellite_image(req.lat, req.lon)
        
        # 2. Process the image to calculate NDVI
        # This is a simplified example. Real implementation requires band math.
        # Assuming get_satellite_image returns an image that can be decoded by OpenCV
        nparr = np.frombuffer(image_data, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        
        # Placeholder for actual NDVI calculation from multi-band data
        ndvi_value = calculate_ndvi(img, img) # Placeholder

        return AnalysisResult(plot_id=1, ndvi=ndvi_value, evi=0.0, soil_quality=0.0, timestamp="2024-01-01T00:00:00Z")
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))