TerraMind-project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
numpy
opencv-python-headless
scikit-learn
google-generativeai
requests
supabase
python-multipart # For file uploads
sentinelhub
pandas